+++ frontend/packages/manager/src/Manager.jsx
@@ -7,37 +7,36 @@ import { ProfileContxtProvider } from "./store/profile-contxt";
 import { InventoryContextProvider } from "./store/inventory-contxt";
 import { GroupContextProvider } from "./store/group-contxt";
 import { ButtonsContextProvider } from "./store/buttons-contx";
-import { InventoryDevicesValidationContxtProvider } from "./store/inventory-devices-contxt";
+import { InventoryDevicesValidationContxtProvider } from "./store/inventory-devices-validation-contxt";
+import { ProfilesValidationContxtProvider } from "./store/profiles-validation-contxt";
 
 function Uncontrolled() {
     return (
-        <TabLayout defaultActivePanelId="one">
-            <TabLayout.Panel label="Profiles" panelId="one">
-                <ButtonsContextProvider>
+        <ButtonsContextProvider>
+            <TabLayout defaultActivePanelId="one">
+                <TabLayout.Panel label="Profiles" panelId="one">
                     <ProfileContxtProvider>
-                        <ProfilesPage />
+                        <ProfilesValidationContxtProvider>
+                            <ProfilesPage />
+                        </ProfilesValidationContxtProvider>
                     </ProfileContxtProvider>
-                </ButtonsContextProvider>
-            </TabLayout.Panel>
-            <TabLayout.Panel label="Inventory" panelId="two">
-                <ButtonsContextProvider>
+                </TabLayout.Panel>
+                <TabLayout.Panel label="Inventory" panelId="two">
                     <InventoryContextProvider>
                         <InventoryDevicesValidationContxtProvider>
                             <InventoryPage />
                         </InventoryDevicesValidationContxtProvider>
                     </InventoryContextProvider>
-                </ButtonsContextProvider>
-            </TabLayout.Panel>
-            <TabLayout.Panel label="Groups" panelId="three">
-                <ButtonsContextProvider>
+                </TabLayout.Panel>
+                <TabLayout.Panel label="Groups" panelId="three">
                     <GroupContextProvider>
                         <InventoryDevicesValidationContxtProvider>
                             <GroupsPage />
                         </InventoryDevicesValidationContxtProvider>
                     </GroupContextProvider>
-                </ButtonsContextProvider>
-            </TabLayout.Panel>
-        </TabLayout>
+                </TabLayout.Panel>
+            </TabLayout>
+        </ButtonsContextProvider>
     );
 }
