+++ frontend/packages/manager/src/components/validation/ValidateInventoryAndGroup.jsx
