+++ frontend/packages/manager/src/store/inventory-devices-validation-contxt.jsx
